Address comments & failing test
